third_party/WebKit/LayoutTests/fast/css-intrinsic-dimensions/fitcontent-minmax-content-inlinesize-contribution-nonreplaced-blocks.html

<!DOCTYPE html>

<style>
@import "resources/width-keyword-classes.css";
@import "resources/intrinsic-size-contribution.css";
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<body onload="checkLayout('.container')">
<div id=log></div>

<div class="container min-content" data-expected-width="60" data-expected-height="40" data-expected-client-width="50" data-expected-client-height="30">
    <div class="item fit-content" data-expected-width="50" data-expected-height="30">XXX XXXXX</div>
</div>

<div class="container min-content" data-expected-width="70" data-expected-height="50" data-expected-client-width="60" data-expected-client-height="40">
    <div class="item fit-content border" data-expected-width="60" data-expected-height="40" data-expected-client-width="50" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container min-content" data-expected-width="80" data-expected-height="60" data-expected-client-width="70" data-expected-client-height="50">
    <div class="item fit-content padding" data-expected-width="70" data-expected-height="50" data-expected-client-width="70" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container min-content" data-expected-width="100" data-expected-height="80" data-expected-client-width="90" data-expected-client-height="70">
    <div class="item fit-content margin" data-expected-width="50" data-expected-height="30" data-expected-client-width="50" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container min-content" data-expected-width="90" data-expected-height="70" data-expected-client-width="80" data-expected-client-height="60">
    <div class="item fit-content border padding" data-expected-width="80" data-expected-height="60" data-expected-client-width="70" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container min-content" data-expected-width="110" data-expected-height="90" data-expected-client-width="100" data-expected-client-height="80">
    <div class="item fit-content border margin" data-expected-width="60" data-expected-height="40" data-expected-client-width="50" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container min-content" data-expected-width="130" data-expected-height="110" data-expected-client-width="120" data-expected-client-height="100">
    <div class="item fit-content border padding margin" data-expected-width="80" data-expected-height="60" data-expected-client-width="70" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="100" data-expected-height="40" data-expected-client-width="90" data-expected-client-height="30">
    <div class="item fit-content" data-expected-width="90" data-expected-height="30">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="110" data-expected-height="50" data-expected-client-width="100" data-expected-client-height="40">
    <div class="item fit-content border" data-expected-width="100" data-expected-height="40" data-expected-client-width="90" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="120" data-expected-height="60" data-expected-client-width="110" data-expected-client-height="50">
    <div class="item fit-content padding" data-expected-width="110" data-expected-height="50" data-expected-client-width="110" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="140" data-expected-height="80" data-expected-client-width="130" data-expected-client-height="70">
    <div class="item fit-content margin" data-expected-width="90" data-expected-height="30" data-expected-client-width="90" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="130" data-expected-height="70" data-expected-client-width="120" data-expected-client-height="60">
    <div class="item fit-content border padding" data-expected-width="120" data-expected-height="60" data-expected-client-width="110" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="150" data-expected-height="90" data-expected-client-width="140" data-expected-client-height="80">
    <div class="item fit-content border margin" data-expected-width="100" data-expected-height="40" data-expected-client-width="90" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container max-content" data-expected-width="170" data-expected-height="110" data-expected-client-width="160" data-expected-client-height="100">
    <div class="item fit-content border padding margin" data-expected-width="120" data-expected-height="60" data-expected-client-width="110" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="100" data-expected-height="40" data-expected-client-width="90" data-expected-client-height="30">
    <div class="item fit-content" data-expected-width="90" data-expected-height="30">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="110" data-expected-height="50" data-expected-client-width="100" data-expected-client-height="40">
    <div class="item fit-content border" data-expected-width="100" data-expected-height="40" data-expected-client-width="90" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="120" data-expected-height="60" data-expected-client-width="110" data-expected-client-height="50">
    <div class="item fit-content padding" data-expected-width="110" data-expected-height="50" data-expected-client-width="110" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="140" data-expected-height="80" data-expected-client-width="130" data-expected-client-height="70">
    <div class="item fit-content margin" data-expected-width="90" data-expected-height="30" data-expected-client-width="90" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="130" data-expected-height="70" data-expected-client-width="120" data-expected-client-height="60">
    <div class="item fit-content border padding" data-expected-width="120" data-expected-height="60" data-expected-client-width="110" data-expected-client-height="50">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="150" data-expected-height="90" data-expected-client-width="140" data-expected-client-height="80">
    <div class="item fit-content border margin" data-expected-width="100" data-expected-height="40" data-expected-client-width="90" data-expected-client-height="30">XXX XXXXX</div>
</div>

<div class="container fit-content" data-expected-width="170" data-expected-height="110" data-expected-client-width="160" data-expected-client-height="100">
    <div class="item fit-content border padding margin" data-expected-width="120" data-expected-height="60" data-expected-client-width="110" data-expected-client-height="50">XXX XXXXX</div>
</div>

</body>
